cav_box/faust/requirements.txt
confluent-kafka==1.5.0
requests==2.25.1
avro-python3==1.10.1
fastavro==1.2.1
faust
aiohttp
aiohttp-sse